% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BE_POP}
\alias{BE_POP}
\alias{SPMA}
\alias{BE_POP_MUNTY}
\alias{BE_POP_COMTY}
\alias{BE_POP_ARRD}
\alias{BE_POP_PROV}
\alias{BE_POP_RGN}
\alias{SPMA_MUNTY}
\alias{SPMA_COMTY}
\alias{SPMA_ARRD}
\alias{SPMA_PROV}
\alias{SPMA_RGN}
\title{Population datasets by location, sex and age within Belgium extracted from Open Data at Statistics Belgium.}
\source{
\url{https://statbel.fgov.be/en/themes/population}
}
\description{
Population datasets by location, sex and age within Belgium extracted from Open Data at Statistics Belgium. Namely:

\itemize{
\item BE_POP_MUNTY: data.frame with data at the level of the municipality
\item BE_POP_COMTY: data.frame with data at the level of the community
\item BE_POP_ARRD: data.frame with data at the level of the district
\item BE_POP_PROV: data.frame with data at the level of the province
\item BE_POP_RGN: data.frame with data at the level of the region
\item BE_POP: data.frame with data at the level of the whole of Belgium
}

Mark that Brussels is not considered as a province but for convenience, the province level of Brussels is set to the region information.

The data contains the following elements which were available at different levels.
\itemize{
\item REFNIS: NIS-code for then region
\item DESCR_DE: description of the region in German
\item DESCR_NL: description of the region in Dutch
\item DESCR_EN: description of the region in English
\item DESCR_FR: description of the region in French
\item YEAR: Population reference year on the 1st of January
\item SEX: Sex of the population subgroup ("M" for men, "F" for women and "MF" for both)
\item AGE: Age of the population subgroup (exact age)
\item AGE5: Age of the population subgroup (by 5-year age band)
\item AGE10: Age of the population subgroup (by 10-year age band)
\item POPULATION: Number of inhabitants in Belgium on the 1st of January
}
}
\examples{
\dontrun{
data(BE_POP_MUNTY)
class(BE_POP_MUNTY)
str(BE_POP_MUNTY)
}
}
\references{
\url{http://statbel.fgov.be/nl/statistieken/opendata/datasets/tools}
}
