% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BE_POP_PROJ}
\alias{BE_POP_PROJ}
\alias{BE_POP_PROJ_ARRD}
\alias{BE_POP_PROJ_PROV}
\alias{BE_POP_PROJ_RGN}
\title{Population projections by location, sex and age within Belgium extracted from Open Data at Planbureau Belgium.}
\source{
\url{https://www.plan.be/databases/}
}
\description{
Population projections by location, sex and age within Belgium extracted from Open Data at Planbureau Belgium.

\itemize{
#' \item BE_POP_PROJ_ARRD: data.frame with data at the level of the district
\item BE_POP_PROJ_PROV: data.frame with data at the level of the province
\item BE_POP_PROJ_RGN: data.frame with data at the level of the region
\item BE_POP_PROJ: data.frame with data at the level of the whole of Belgium
}

The data contains the following elements which were available at different levels.
\itemize{
\item REGION: Name of the region
\item YEAR: Population reference year on the 1st of January
\item SEX: Sex of the population subgroup ("M" for men, "F" for women and "MF" for both)
\item AGE: Age of the population subgroup (exact age)
\item AGE5: Age of the population subgroup (by 5-year age band)
\item AGE10: Age of the population subgroup (by 10-year age band)
\item POPULATION: Number of inhabitants in Belgium on the 1st of January
}
}
\examples{
\dontrun{
data(BE_POP_PROJ)
class(BE_POP_PROJ)
str(BE_POP_PROJ)
}
}
\references{
\url{https://www.plan.be/databases/}
}
