% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bopr.R
\name{BOPR}
\alias{BOPR}
\title{BOPR}
\usage{
BOPR(x, ...)
}
\arguments{
\item{x}{a  matrix of predictors.}

\item{y}{a factor vector with 2 level}

\item{beta}{scaling parameter}

\item{prior_prob}{prior of initial parametes}

\item{epsilon}{parameter to apply dynamics}

\item{formula}{an optional data frame in which to interpret the variables named in the formula.}

\item{data}{an optional data frame in which to interpret the variables named in the formula.}

\item{subset}{optional expression saying that only a subset of the rows of the data should be used in the fit.(currently it's not working.)}

\item{na.action}{a function which indicates what should happen when the data contain \code{NA}s.}
}
\value{
S3 \code{BOPR} object; a list of consisting of
 \item{beta_matrix}{beta matrix with mean and variance}
 \item{beta}{scaling parameter}
 \item{prior_prob}{prior of initial parametes}
 \item{epsilon}{parameter to apply dynamics}
}
\description{
Bayesian online learning scheme for probit regression (BOPR)
}
\author{
Original Python code by Andred Tulloch, R code and modifications by Heewon Jeon
}
\references{
Graepel, Thore, et al. "Web-scale bayesian click-through rate prediction for sponsored search advertising in microsoft's bing search engine." Proceedings of the 27th International Conference on Machine Learning (ICML-10). 2010.
He, X., Bowers, S., Candela, J. Q., Pan, J., Jin, O., Xu, T., … Herbrich, R. (2014). Practical Lessons from Predicting Clicks on Ads at Facebook. Proceedings of 20th ACM SIGKDD Conference on Knowledge Discovery and Data Mining - ADKDD’14, 1–9.
}

