% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit_approval.R
\docType{data}
\name{credit_approval}
\alias{credit_approval}
\title{Credit Approval Data Set}
\format{A data frame with 690 rows and 15 variables:
\describe{
  \item{A1}{b,a}
  \item{A2}{continuous}
  \item{A3}{continuous}
  \item{A4}{u, y, l, t. }
  \item{A5}{g, p, gg. }
  \item{A6}{c, d, cc, i, j, k, m, r, q, w, x, e, aa, ff. }
  \item{A7}{v, h, bb, j, n, z, dd, ff, o. }
  \item{A8}{continuous}
  \item{A9}{t,f}
  \item{A10}{t,f}
  \item{A11}{continuous}
  \item{A12}{t,f}
  \item{A13}{g, p, s.}
  \item{A14}{continuous}
  \item{A15}{continuous}
  \item{A16}{1,-1 ; class}
}}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Credit+Approval}
}
\usage{
credit_approval
}
\description{
This data concerns credit card applications. All attribute names and values have been changed to meaningless symbols to protect confidentiality of the data. This dataset is interesting because there is a good mix of attributes -- continuous, nominal with small numbers of values, and nominal with larger numbers of values. There are also a few missing values.
}
\keyword{datasets}

