% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FPB}
\alias{FPB}
\title{Population projection for Belgium}
\source{
\url{https://www.plan.be/databases/data-35-en-population_projections_2021_2070}
}
\description{
Population projection for Belgium by ages and sex
}
\examples{
\dontrun{
data(BE_POP_MUNTY)
class(BE_POP_MUNTY)
str(BE_POP_MUNTY)
}
}
\references{
\url{https://www.plan.be/databases/data-35-en-population_projections_2021_2070}
}
