% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bopr.R
\name{predict.BOPR}
\alias{predict.BOPR}
\title{Predict new samples using a BOPR model}
\usage{
\method{predict}{BOPR}(object, newdata = NULL, type = "response",
  na.action = na.pass, ...)
}
\arguments{
\item{object}{\code{BOPR} object}

\item{newdata}{a data.frame of predictors}

\item{type}{either "class" for the predicted class or "prob" for model confidence values.}

\item{na.action}{when using a formula for the original model fit, how should missing values be handled?}

\item{...}{other options (not currently used)}
}
\value{
when type = "class", a factor vector is returned. When type = "response", probability is returned.
}
\description{
This function produces predicted values from a BOPR model.
}

