% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{selectData}
\alias{selectData}
\title{Extract data from the Belgian Federal Plan Bureau}
\usage{
selectData(
  arr,
  gender,
  year = NULL,
  ageGroup = NULL,
  ARROND = ARROND,
  FPB = FPB
)
}
\arguments{
\item{arr}{Name of the arrondissement. The list of arrondissement is provided
in the object ARROND}

\item{gender}{Gender for which you want the data - should be in "MF" -
Male and Female, "F"- female or "M"-male}

\item{year}{Year for which the data is requested}

\item{ageGroup}{Age fro whihc the data is requested - from 0 to 110
\itemize{
\item optionnal
}}
}
\value{
a df (if only arr and gender provide), a vector (if three parameters are provided)
or a single value ( if all four parameters are provided)
}
\description{
\code{selectData} extrart data from a list based on four parameters: arrondissement,
gender, year and age
}
